Separate TaskAdapter into it's own file
